#4033: Never let chrome remembrer the user password
After #199 we made our login forms chrome-password agent friendly. For some odd reason it was keeping the user's password in the DOM for no reason.  This commit don't render the login unless a given user is about to login.

This commit also cleans up the user's password after login (The live site had this issue)