#9541: Mobile web - Support for password protected links
Mobile client depends on order of logical operation while checking the needed page to parse and finally it always failback to start page.
While, this is not the same logic we have for desktop webclient.
And due to the nature of absence of some html elements in Start page in mobile [like dialog div]